requirements.txt
streamlit==1.49.1
python-dotenv==1.1.1
langchain==0.3.27
langchain-community==0.3.29
faiss-cpu==1.12.0
openai==1.102.0
tiktoken==0.11.0
pypdf==6.0.0
